+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/atomist/cli/compare/0.2.1...HEAD)
+## [Unreleased](https://github.com/atomist/cli/compare/0.3.0...HEAD)
+
+## [0.3.0](https://github.com/atomist/cli/compare/0.2.1...0.3.0) - 2018-08-01
 
 ### Added